+++ modules/example/src/main/scala/zio/redis/example/ApiError.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package example
+package zio.redis.example
 
 import zhttp.http._
