+++ modules/example/src/main/scala/zio/redis/example/Contributor.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package example
+package zio.redis.example
 
 import zio.json.{DeriveJsonCodec, JsonCodec}
