+++ modules/example/src/main/scala/zio/redis/example/package.scala
@@ -14,8 +14,17 @@
  * limitations under the License.
  */
 
-package example
+package zio.redis
 
-import zio.prelude.Newtype
+import sttp.capabilities.WebSockets
+import sttp.capabilities.zio.ZioStreams
+import sttp.client3.SttpBackend
+import zio.Task
 
-object Owner extends Newtype[String]
+package object example {
+  type Contributions = Contributions.Type
+  type Login         = Login.Type
+  type Owner         = Owner.Type
+  type Name          = Name.Type
+  type Sttp          = SttpBackend[Task, ZioStreams with WebSockets]
+}
